mxt224e: T48 threshold entry
This is the real one.

	cd /sys/kernel/mxt224e

	# check
	cat threshold_t48

	# threshold value in config t48: index 35
	cat config_t48 | grep 35

	# on
	echo on > threshold_t48

	# off
	echo off > threshold_t48

	# val
	echo val=?? > threshold_t48

It's the same:
	echo 35 ?? > config-t48

But threshold_t48 entry will apply user's value on waking.

When the threshold value is too big, touchscreen will lose responses.
When the threshold value is low (<10), air touches will appear!
When the threshold value is too low (<7), mxT224E calibration will fail, and the points may fly here and there on the screen (signal is too noisy).

Signed-off-by: cocafe <cocafehj@gmail.com>